\name{tiny-package}
\alias{tiny-package}
\alias{tiny}
\docType{package}
\title{
\packageTitle{tiny}
}
\description{
\packageDescription{tiny}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tiny}
\packageIndices{tiny}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{tiny}

Maintainer: \packageMaintainer{tiny}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
